Split IfTest
